backend/requirements.txt
fastapi[all]
uvicorn
pymongo
passlib[bcrypt]
fastapi-jwt-auth[asymmetric]
reportlab
matplotlib>=3.2.2
pandas
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  
torchvision
tqdm
seaborn

